Divide utility functions into public and private
