+++ taxcalc/filings/forms/tax_form.py
@@ -1,7 +1,7 @@
 import six
 
 from taxcalc.filings.forms.errors import UnsupportedFormYearError
-from taxcalc.utils import string_to_number
+from taxcalc.filings.forms.utils import string_to_number
 
 
 class TaxForm(object):
